tdmg working
