+++ tests/_data/tut2_01/c01/dmg2/attr02_tut2c_Tut2c.csv
@@ -1,0 +1,35 @@
+rEventName,haz_1000yr_cT3,haz_1000yr_cT3,haz_1000yr_fail_A_cT4,haz_1000yr_fail_A_cT4,haz_1000yr_fail_B_cT4,haz_1000yr_fail_B_cT4,haz_100yr_cT2,haz_100yr_cT2,haz_200yr_cT2,haz_200yr_cT2,haz_50yr_cT3,haz_50yr_cT3
+nestID,f0,f1,f0,f1,f0,f1,f0,f1,f0,f1,f0,f1
+xid,,,,,,,,,,,,
+14879,0.8203054806828391,0.17969451931716082,0.8203054806828391,0.17969451931716082,0.8203054806828391,0.17969451931716082,0.6479671511108044,0.35203284888919556,0.6519950215563857,0.3480049784436144,0.0,1.0
+14880,0.8698769514813832,0.13012304851861675,0.8698769514813832,0.13012304851861675,0.8698769514813832,0.13012304851861675,0.6923008524491399,0.3076991475508601,0.699813276695289,0.30018672330471097,0.0,1.0
+14925,0.819410787173293,0.18058921282670706,0.8424258750831755,0.1575741249168245,0.8424258750831755,0.1575741249168245,0.6368385195213823,0.3631614804786178,0.6372524893247792,0.3627475106752208,0.0,1.0
+14926,0.7823408405930518,0.21765915940694824,0.8142344557337056,0.18576554426629446,0.8142344557337056,0.18576554426629446,0.5771175548229858,0.4228824451770143,0.589776575914306,0.41022342408569407,0.0,1.0
+14927,0.8582609875421996,0.14173901245780038,0.8765539999713605,0.12344600002863947,0.8765539999713605,0.12344600002863947,0.6996167109231379,0.30038328907686207,0.7005444604345423,0.2994555395654577,0.0,1.0
+14928,0.8660826701819754,0.13391732981802457,0.869281045751634,0.13071895424836602,0.869281045751634,0.13071895424836602,0.7086571123637818,0.29134288763621824,0.7215505402476419,0.2784494597523582,0.0,1.0
+14933,0.832684766031205,0.16731523396879497,0.8555068729311835,0.14449312706881645,0.8555068729311835,0.14449312706881645,0.6579750084103324,0.34202499158966765,0.6601002568282459,0.3398997431717541,0.0,1.0
+14934,0.848237687462879,0.15176231253712105,0.848237687462879,0.15176231253712105,0.848237687462879,0.15176231253712105,0.6800176382439312,0.3199823617560687,0.8203917805998375,0.17960821940016247,0.5,0.5
+14935,0.8262108032369837,0.17378919676301632,0.8262108032369837,0.17378919676301632,0.8262108032369837,0.17378919676301632,0.632981718153428,0.367018281846572,0.7756896630190507,0.2243103369809492,0.5,0.5
+14936,0.849624060150376,0.15037593984962405,0.849624060150376,0.15037593984962405,0.849624060150376,0.15037593984962405,0.677631007706791,0.3223689922932091,0.69586885495906,0.30413114504094,0.5,0.5
+14937,0.8239111971488828,0.1760888028511172,0.8239111971488828,0.1760888028511172,0.8239111971488828,0.1760888028511172,0.636749818632425,0.3632501813675751,0.7924063641698391,0.2075936358301609,0.5,0.5
+14938,0.8529411764705882,0.14705882352941177,0.8529411764705882,0.14705882352941177,0.8529411764705882,0.14705882352941177,0.7616018960043579,0.238398103995642,0.8375574364885077,0.16244256351149236,0.5,0.5
+14939,0.8414395965411506,0.1585604034588493,0.8414395965411506,0.1585604034588493,0.8414395965411506,0.1585604034588493,0.6661684776065111,0.3338315223934889,0.812417879031156,0.1875821209688439,0.0,1.0
+14940,0.8413199337961376,0.1586800662038625,0.8413199337961376,0.1586800662038625,0.8413199337961376,0.1586800662038625,0.6593902480956971,0.34060975190430287,0.8084170400489427,0.1915829599510574,0.5,0.5
+14941,0.6996996996996997,0.3003003003003003,0.6996996996996997,0.3003003003003003,0.6996996996996997,0.3003003003003003,0.6996996996996997,0.3003003003003003,0.6996996996996997,0.3003003003003003,0.5,0.5
+14942,0.8408756067970048,0.15912439320299518,0.8408756067970048,0.15912439320299518,0.8408756067970048,0.15912439320299518,0.6584197818135445,0.34158021818645545,0.8071816837370056,0.19281831626299437,0.5,0.5
+14943,0.8092286189731601,0.1907713810268399,0.8092286189731601,0.1907713810268399,0.8092286189731601,0.1907713810268399,0.7758719011826362,0.2241280988173639,0.7829572131816108,0.21704278681838912,0.5,0.5
+14944,0.8333333333333334,0.16666666666666666,0.8333333333333334,0.16666666666666666,0.8333333333333334,0.16666666666666666,0.8333333333333334,0.16666666666666666,0.8333333333333334,0.16666666666666666,0.7869122924578856,0.21308770754211434
+14945,0.5,0.5,0.6183539655538809,0.3816460344461191,0.587651647254599,0.412348352745401,0.5,0.5,0.5,0.5,0.5,0.5
+14946,0.835084548523832,0.1649154514761681,0.835084548523832,0.1649154514761681,0.835084548523832,0.1649154514761681,0.6527611455051788,0.34723885449482117,0.8049423832269207,0.19505761677307934,0.5,0.5
+14947,0.8095238095238095,0.19047619047619047,0.8095238095238095,0.19047619047619047,0.8095238095238095,0.19047619047619047,0.6368270942886344,0.36317290571136557,0.7827704487600643,0.21722955123993565,0.5,0.5
+14948,0.8559621998720368,0.1440378001279632,0.8559621998720368,0.1440378001279632,0.8559621998720368,0.1440378001279632,0.6925399377040066,0.30746006229599343,0.8291976429453346,0.17080235705466534,0.5,0.5
+14955,0.8387096774193549,0.16129032258064516,0.8387096774193549,0.16129032258064516,0.8387096774193549,0.16129032258064516,0.8223644615271362,0.17763553847286387,0.827449651683846,0.17255034831615396,0.5,0.5
+15152,0.6759145391472493,0.32408546085275064,0.8469977179602379,0.1530022820397621,0.8406576275529298,0.15934237244707014,0.0,1.0,0.0,1.0,0.0,1.0
+15155,0.6543079452479096,0.3456920547520904,0.8300981242872026,0.16990187571279738,0.8253109526426509,0.17468904735734914,0.0,1.0,0.0,1.0,0.0,1.0
+17873,0.0,1.0,0.8631664398726426,0.13683356012735737,0.8598582702488119,0.14014172975118813,0.5,0.5,0.0,1.0,0.5,0.5
+17899,0.9282296650717703,0.07177033492822966,0.9282296650717703,0.07177033492822966,0.9282296650717703,0.07177033492822966,0.9282296650717703,0.07177033492822966,0.9282296650717703,0.07177033492822966,0.8448658806654954,0.1551341193345046
+17900,0.8351648351648352,0.16483516483516483,0.8351648351648352,0.16483516483516483,0.8351648351648352,0.16483516483516483,0.705233802633185,0.294766197366815,0.7698412309124933,0.2301587690875067,0.0,1.0
+17907,0.8935293280355708,0.1064706719644293,0.8970066945785848,0.10299330542141516,0.8970066945785848,0.10299330542141516,0.7385600215365749,0.2614399784634251,0.7467636335443085,0.2532363664556915,0.5,0.5
+71528,0.7777028556485095,0.2222971443514906,0.9006189471845556,0.09938105281544435,0.9006189471845556,0.09938105281544435,0.0,1.0,0.0,1.0,0.0,1.0
+74651,0.9029222719535603,0.09707772804643972,0.9029222719535603,0.09707772804643972,0.9029222719535603,0.09707772804643972,0.9027317328462059,0.09726826715379408,0.9029222719535603,0.09707772804643972,0.7488637188449273,0.2511362811550727
+75511,0.5,0.5,0.5,0.5,0.5,0.5,0.5,0.5,0.5,0.5,0.5,0.5
