+++ tests/_data/tut2_01/c01/dmg2/dmgs_tut2c_Tut2c.csv
@@ -1,0 +1,33 @@
+xid,haz_1000yr_cT3,haz_1000yr_fail_A_cT4,haz_1000yr_fail_B_cT4,haz_100yr_cT2,haz_200yr_cT2,haz_50yr_cT3
+14879,111300.0,111300.0,111300.0,56812.88,57470.442203,20000.0
+14880,153700.67200000002,153700.67200000002,153700.67200000002,64998.555112,66625.19840925334,20000.0
+14925,110748.58618046,126924.392,126924.392,55071.92,55134.768431,20000.0
+14926,91886.78323711999,107662.592,107662.592,47294.467358719994,48753.92,20000.0
+14927,141104.4119272,162014.16,162014.16,66581.6,66787.87785666667,20000.0
+14928,149345.86903112,153000.0,153000.0,68647.63427818,71826.32,20000.0
+14933,119534.8416614,138414.888,138414.888,58475.259094500005,58840.88,20000.0
+14934,131785.024,131785.024,131785.024,62503.445159413335,111353.47851448,0.0
+14935,115081.952,115081.952,115081.952,54493.19826624,89162.18605519999,0.0
+14936,133000.0,133000.0,133000.0,62040.70638968,65761.10446468,0.0
+14937,113579.056,113579.056,113579.056,55058.472165666666,96342.06713524999,0.0
+14938,136000.0,136000.0,136000.0,83893.28465618001,123120.44065092,0.0
+14939,126134.89600000001,126134.89600000001,126134.89600000001,59910.45979302667,106619.96941234,20000.0
+14940,126039.77600000001,126039.77600000001,126039.77600000001,58718.22485464,104393.4178964,0.0
+14941,66600.0,66600.0,66600.0,66600.0,66600.0,0.0
+14942,125687.83200000001,125687.83200000001,125687.83200000001,58551.4000377,103724.58585688,0.0
+14943,104837.528,104837.528,104837.528,89234.683672115,92147.72945546,0.0
+14944,120000.0,120000.0,120000.0,120000.0,120000.0,93858.06544493999
+14945,0.0,52404.57962317333,48502.679510760005,0.0,0.0,0.0
+14946,121274.26400000001,121274.26400000001,121274.26400000001,57597.24103772001,102533.80683547999,0.0
+14947,105000.0,105000.0,105000.0,55070.18746573334,92068.50488729999,0.0
+14948,138852.44,138852.44,138852.44,65049.0988997,117094.40282255001,0.0
+14955,124000.0,124000.0,124000.0,112590.08288510499,115908.19836164668,0.0
+15152,61712.117375999995,130717.00456599999,125515.89193039999,20000.0,20000.0,20000.0
+15155,57854.96,117715.00412278001,114489.146873,20000.0,20000.0,20000.0
+17873,15000.0,109622.23000000001,107034.50019228001,0.0,15000.0,0.0
+17899,209000.0,209000.0,209000.0,209000.0,209000.0,96690.52858486
+17900,91000.0,91000.0,91000.0,50887.788810239996,65172.40276992,15000.0
+17907,140883.86710859998,145640.534,145640.534,57374.545730000005,59233.19865128667,0.0
+71528,67477.25007336299,150934.203,150934.203,15000.0,15000.0,15000.0
+74651,154515.359,154515.359,154515.359,154212.67838855402,154515.359,59728.526404107004
+75511,0.0,0.0,0.0,0.0,0.0,0.0
