+++ tests/_data/tut2_01/c01/exlikes_tut2c.csv
@@ -1,0 +1,33 @@
+xid,haz_1000yr_fail_A_cT4,haz_1000yr_fail_B_cT4
+14879,0.05,0.0
+14880,0.15,0.0
+14925,0.15,0.1
+14926,0.15,0.1
+14927,0.15,0.1
+14928,0.05,0.1
+14933,0.15,0.1
+14934,0.15,0.1
+14935,0.15,0.1
+14936,0.15,0.1
+14937,0.15,0.1
+14938,0.15,0.1
+14939,0.15,0.1
+14940,0.15,0.1
+14941,0.15,0.0
+14942,0.15,0.1
+14943,0.15,0.1
+14944,0.15,0.1
+14945,0.15,0.1
+14946,0.15,0.1
+14947,0.15,0.1
+14948,0.15,0.1
+14955,0.15,0.0
+15152,0.15,0.1
+15155,0.15,0.0
+17873,0.15,0.1
+17899,0.15,0.0
+17900,0.05,0.0
+17907,0.15,0.1
+71528,0.15,0.1
+74651,0.05,0.1
+75511,0.0,0.0
