+++ tests/_data/tut2_01/c01/expos_tut2c_6_32.csv
@@ -1,0 +1,33 @@
+xid,haz_1000yr_cT3,haz_1000yr_fail_A_cT4,haz_1000yr_fail_B_cT4,haz_100yr_cT2,haz_200yr_cT2,haz_50yr_cT3
+14879,13.716447830200195,15.096734046936035,14.896734237670898,12.076109886169434,12.372537612915039,10.037759780883789
+14880,13.711297035217285,15.091536521911621,14.891536712646484,12.070693016052246,12.367313385009766,10.031964302062988
+14925,13.753494262695312,15.13412094116211,14.934121131896973,12.115073204040527,12.410107612609863,10.07548999786377
+14926,13.735912322998047,15.112829208374023,14.912829399108887,12.092883110046387,12.388710975646973,10.055706024169922
+14927,13.755128860473633,15.137419700622559,14.937419891357422,12.118510246276855,12.413421630859375,10.0773286819458
+14928,13.735664367675781,15.119675636291504,14.919675827026367,12.100018501281738,12.395591735839844,10.061222076416016
+14933,13.737154006958008,15.114081382751465,14.914081573486328,12.094188690185547,12.389969825744629,10.05710220336914
+14934,13.727742195129395,15.11168098449707,14.911681175231934,12.091686248779297,12.387557029724121,10.048629760742188
+14935,13.731010437011719,15.107881546020508,14.907881736755371,12.087726593017578,12.383739471435547,10.050189018249512
+14936,13.740422248840332,15.120928764343262,14.920928955078125,12.101324081420898,12.3968505859375,10.06078052520752
+14937,13.730761528015137,15.11117935180664,14.911179542541504,12.089446067810059,12.385396003723145,10.048070907592773
+14938,13.74532413482666,15.125875473022461,14.925875663757324,12.10647964477539,12.401822090148926,10.064457893371582
+14939,13.750226974487305,15.132472038269043,14.932472229003906,12.111635208129883,12.406793594360352,10.071812629699707
+14940,13.748592376708984,15.12917423248291,14.929174423217773,12.109916687011719,12.405136108398438,10.068135261535645
+14941,13.705009460449219,15.085190773010254,14.885190963745117,12.064079284667969,12.360937118530273,10.024888038635254
+14942,13.69847297668457,15.075296401977539,14.875296592712402,12.060903549194336,12.357874870300293,10.011737823486328
+14943,13.717192649841309,15.09748649597168,14.897486686706543,12.076892852783203,12.373292922973633,10.02884578704834
+14944,13.709022521972656,15.082393646240234,14.882393836975098,12.06460189819336,12.361440658569336,10.019651412963867
+14945,13.719572067260742,15.100137710571289,14.900137901306152,12.083093643188477,12.379271507263184,10.039436340332031
+14946,13.716054916381836,15.093040466308594,14.893040657043457,12.07569694519043,12.372138977050781,10.031522750854492
+14947,13.726604461669922,15.107234954833984,14.907235145568848,12.090490341186523,12.386404037475586,10.04734992980957
+14948,13.722591400146484,15.102934837341309,14.902935028076172,12.082571029663086,12.378767967224121,10.040716171264648
+14955,13.697976112365723,15.078093528747559,14.878093719482422,12.056682586669922,12.353804588317871,10.016974449157715
+15152,13.687969207763672,15.0678071975708,14.867807388305664,12.046916007995605,12.344066619873047,10.00792121887207
+15155,13.706395149230957,15.091787338256836,14.8917875289917,12.069235801696777,12.3659086227417,10.026448249816895
+17873,13.82020378112793,15.203341484069824,15.003341674804688,,12.481325149536133,
+17899,13.378049850463867,14.745282173156738,14.545282363891602,11.760359764099121,12.051095008850098,9.738381385803223
+17900,13.348488807678223,14.718463897705078,14.518464088439941,11.736531257629395,12.026734352111816,9.712799072265625
+17907,13.88474178314209,15.266571044921875,15.066571235656738,12.253107070922852,12.543209075927734,10.227132797241211
+71528,13.853734016418457,15.23717975616455,15.037179946899414,12.2224760055542,12.513672828674316,10.19040298461914
+74651,13.798359870910645,15.175848960876465,14.975849151611328,12.158559799194336,12.452040672302246,10.120179176330566
+75511,,,,,,
