+++ tests/_data/tut2_01/c01/finv_tut2c_finv_cT2.csv
@@ -1,0 +1,33 @@
+xid,f0_tag,f0_scale,f0_cap,f0_elv,f1_cap,f1_elv,f1_scale,f1_tag
+14879,BA_S,117.99,91300.0,3.0,20000.0,1.0,117.99,BA_C
+14880,BA_S,140.56,134000.0,3.0,20000.0,-2.0,140.56,BA_C
+14925,BA_S,112.41,138000.0,3.0,20000.0,1.0,112.41,BA_C
+14926,BA_S,92.16,93000.0,3.0,20000.0,1.0,92.16,BA_C
+14927,BA_S,149.3,177000.0,3.0,20000.0,1.0,149.3,BA_C
+14928,BA_S,166.11,133000.0,3.0,20000.0,1.0,166.11,BA_C
+14933,BA_S,124.49,153000.0,3.0,20000.0,1.0,124.49,BA_C
+14934,BA_S,117.52,147000.0,2.0,20000.0,3.0,117.52,BA_C
+14935,BA_S,99.96,96000.0,2.0,20000.0,3.0,99.96,BA_C
+14936,BA_S,127.46,113000.0,2.0,20000.0,3.0,127.46,BA_C
+14937,BA_S,98.38,138000.0,2.0,20000.0,3.0,98.38,BA_C
+14938,BA_S,131.46,116000.0,2.0,20000.0,3.0,131.46,BA_C
+14939,BA_S,111.58,108000.0,2.0,20000.0,0.2,111.58,BA_C
+14940,BA_S,111.48,159000.0,2.0,20000.0,3.0,111.48,BA_C
+14941,BA_S,102.56,46600.0,2.0,20000.0,3.0,102.56,BA_C
+14942,BA_S,111.11,113000.0,2.0,20000.0,2.0,111.11,BA_C
+14943,BA_S,89.19,120000.0,2.0,20000.0,2.0,89.19,BA_C
+14944,BA_S,107.02,100000.0,0.0,20000.0,1.0,107.02,BA_C
+14945,BA_S,111.02,110000.0,6.0,20000.0,5.0,111.02,BA_C
+14946,BA_S,106.47,145000.0,2.0,20000.0,2.0,106.47,BA_C
+14947,BA_S,101.45,85000.0,2.0,20000.0,2.0,101.45,BA_C
+14948,BA_S,124.95,130000.0,2.0,20000.0,2.0,124.95,BA_C
+14955,BA_S,119.47,104000.0,2.0,20000.0,2.0,119.47,BA_C
+15152,BA_S,131.64,175000.0,4.0,20000.0,1.0,131.64,BA_C
+15155,BA_S,121.33,153000.0,4.0,20000.0,1.0,121.33,BA_C
+17873,CA_S,265.42,382937.594,4.0,15000.0,1.0,265.42,CA_C
+17899,CA_S,255.58,194000.0,4.0,15000.0,1.0,255.58,CA_C
+17900,CA_S,92.16,76000.0,1.1,15000.0,0.2,92.16,CA_C
+17907,CA_S,117.62,158966.688,3.0,15000.0,3.0,117.62,CA_C
+71528,CA_S,137.619,135934.203,4.0,15000.0,1.0,137.619,CA_C
+74651,CA_S,137.619,139515.359,1.1,15000.0,0.2,137.619,CA_C
+75511,CA_S,137.619,358502.594,3.0,15000.0,-2.0,137.619,CA_C
