+++ tests/_data/tut2_01/c01/gels_tut2c_1_32.csv
@@ -1,0 +1,33 @@
+xid,dtm_cT1
+14879,10.1053466796875
+14880,9.7265625
+14925,10.2933349609375
+14926,10.4501953125
+14927,10.27197265625
+14928,10.4901123046875
+14933,10.3233642578125
+14934,10.1817626953125
+14935,10.30419921875
+14936,10.507080078125
+14937,10.219970703125
+14938,10.078369140625
+14939,10.2320556640625
+14940,10.309326171875
+14941,9.9996337890625
+14942,10.262939453125
+14943,9.9034423828125
+14944,9.9405517578125
+14945,10.5
+14946,10.2308349609375
+14947,10.30224609375
+14948,10.1812744140625
+14955,9.920654296875
+15152,10.4420166015625
+15155,10.6431884765625
+17873,12.264404296875
+17899,7.1693115234375
+17900,10.9041748046875
+17907,10.09033203125
+71528,10.2237548828125
+74651,10.4173583984375
+75511,98.8564453125
