+++ tests/_data/tut2_01/evals_4_tut2c.csv
@@ -1,0 +1,2 @@
+haz_50yr_cT3,haz_200yr_cT2,haz_100yr_cT2,haz_1000yr_cT3,haz_1000yr_fail_A_cT4,haz_1000yr_fail_B_cT4
+50,200,100,1000,1000,1000
